Move directories to match new namespace
