+++ specification/query_rules/_types/QueryRuleset.ts
@@ -21,7 +21,7 @@ import { Id, IndexName, Name } from '@_types/common'
 import { EpochTime, UnitMillis } from '@_types/Time'
 import { InlineScript } from '@_types/Scripting'
 import { UserDefinedValue } from '@spec_utils/UserDefinedValue'
-import { PinnedDoc } from '../../../_types/query_dsl/specialized'
+import { PinnedDoc } from '../../_types/query_dsl/specialized'
 
 export class QueryRuleset {
   /**
