+++ specification/query_rules/query_rule_delete/QueryRuleDeleteRequest.ts
