+++ specification/query_rules/query_rule_delete/QueryRuleDeleteResponse.ts
