+++ specification/query_rules/query_rule_get/QueryRuleGetRequest.ts
