+++ specification/query_rules/query_rule_put/QueryRulePutResponse.ts
