+++ specification/query_rules/query_ruleset_delete/QueryRulesetDeleteRequest.ts
