+++ specification/query_rules/query_ruleset_delete/QueryRulesetDeleteResponse.ts
