+++ specification/query_rules/query_ruleset_get/QueryRulesetGetResponse.ts
