+++ specification/query_rules/query_ruleset_put/QueryRulesetPutRequest.ts
