+++ specification/query_rules/query_ruleset_put/QueryRulesetPutResponse.ts
